Move WorkplaceQuery to where it's used
