+++ emergence_lib/src/units/actions.rs
@@ -1,16 +1,16 @@
 //! What are units currently doing?
 
-use bevy::{ecs::query::WorldQuery, prelude::*};
+use bevy::{
+    ecs::{query::WorldQuery, system::SystemParam},
+    prelude::*,
+};
 use leafwing_abilities::prelude::Pool;
 use rand::{rngs::ThreadRng, seq::SliceRandom, thread_rng};
 
 use crate::{
     asset_management::manifest::Id,
-    crafting::{
-        components::{
-            CraftingState, InputInventory, OutputInventory, StorageInventory, WorkersPresent,
-        },
-        WorkplaceQuery,
+    crafting::components::{
+        CraftingState, InputInventory, OutputInventory, StorageInventory, WorkersPresent,
     },
     items::{
         item_manifest::{Item, ItemManifest},
@@ -1055,6 +1055,57 @@ impl CurrentAction {
             action: UnitAction::Abandon,
             timer: Timer::from_seconds(0.1, TimerMode::Once),
             just_started: true,
+        }
+    }
+}
+
+/// A query about the [`CraftingState`] of a structure that might need work done.
+#[derive(SystemParam)]
+pub(crate) struct WorkplaceQuery<'w, 's> {
+    /// The contained query type.
+    query: Query<
+        'w,
+        's,
+        (
+            &'static CraftingState,
+            &'static Id<Structure>,
+            &'static WorkersPresent,
+        ),
+    >,
+}
+
+impl<'w, 's> WorkplaceQuery<'w, 's> {
+    /// Is there a structure of type `structure_id` at `structure_pos` that needs work done by a unit?
+    ///
+    /// If so, returns `Some(matching_structure_entity_that_needs_work)`.
+    pub(crate) fn needs_work(
+        &self,
+        structure_pos: TilePos,
+        structure_id: Id<Structure>,
+        map_geometry: &MapGeometry,
+    ) -> Option<Entity> {
+        // Prioritize ghosts over structures to allow for replacing structures by building
+        let entity = if let Some(ghost_entity) = map_geometry.get_ghost(structure_pos) {
+            ghost_entity
+        } else {
+            map_geometry.get_structure(structure_pos)?
+        };
+
+        let (found_crafting_state, found_structure_id, workers_present) =
+            self.query.get(entity).ok()?;
+
+        if *found_structure_id != structure_id {
+            return None;
+        }
+
+        if let CraftingState::InProgress { .. } = found_crafting_state {
+            if workers_present.needs_more() {
+                Some(entity)
+            } else {
+                None
+            }
+        } else {
+            None
         }
     }
 }
